FEAT: Finished scheduling - next will beautify the adding exercise section a little bit (right now it just displays text - could be a card in a scrollable list).
Unless other features are required/wanted (e.g. set rest time for an exercise, or specifying reps/sets), the saving of a workout can be complete. Unsure if I want to allow for explicit sets/reps, or if I will let the user decide on the day WITH a reminder of their previous sets/reps (maybe a last 1-3 workouts?)