FEAT: Implemented exercise entries - still need some work on the run workout page (e.g. historical data, saving, displaying the previous and current tags (currently looks quite out of place)).
May also be worth adding something to show the user that this is meant to be a table that they can add to (e.g. have it as a full but empty table initially, rather than just table headers).

Maybe temporarily saving a workout would be nice (rather than doing it all and then completing (what if app breaks or closes mid way through the workout..?!?))